Adjustments to Github Orgs / Repositories

+++ cla-backend/cla/controllers/github.py
@@ -11,6 +11,7 @@ from cla.auth import AuthUser
 from cla.models import DoesNotExist
 from cla.models.dynamo_models import UserPermissions
 from cla.controllers.github_application import GitHubInstallation
+from cla.controllers.project import check_user_authorization
 
 
 def get_organizations():
@@ -40,77 +41,89 @@ def get_organization(organization_name):
     return github_organization.to_dict()
 
 
-def create_organization(organization_name,
-                        organization_project_id=None,
-                        organization_installation_id=None):
+def create_organization(auth_user,
+                        organization_name,
+                        organization_sfid):
     """
     Creates a github organization and returns the newly created github organization in dict format.
 
+    :param auth_user: authorization for this user.
+    :type auth_user: AuthUser
     :param organization_name: The github organization name.
-    :type organization_name: string
-    :param organization_project_id: The ID of the github organization project.
-    :type organization_project_id: string/None
-    :param organization_installation_id: The github app installation id.
-    :type organization_installation_id: string/None
+    :type organization_name: string 
+    :param organization_sfid: The SFDC ID for the github organization. 
+    :type organization_sfid: string/None
     :return: dict representation of the new github organization object.
     :rtype: dict
     """
+    # Validate user is authorized for this SFDC ID. 
+    can_access = check_user_authorization(auth_user, organization_sfid)
+    if not can_access['valid']:
+        return can_access['errors']
+
     github_organization = get_github_organization_instance()
     try:
         github_organization.load(str(organization_name))
     except DoesNotExist as err:
-        github_organization.set_organization_name(organization_name)
-        if organization_project_id:
-            github_organization.set_organization_project_id(str(organization_project_id))
-        if organization_installation_id:
-            github_organization.set_organization_installation_id(organization_installation_id)
+        github_organization.set_organization_name(str(organization_name))
+        github_organization.set_organization_sfid(str(organization_sfid))
         github_organization.save()
         return github_organization.to_dict()
     return {'errors': {'organization_name': 'This organization already exists'}}
 
 
 def update_organization(organization_name, # pylint: disable=too-many-arguments
-                        organization_project_id=None,
+                        organization_sfid=None,
                         organization_installation_id=None):
     """
     Updates a github organization and returns the newly updated org in dict format.
     Values of None means the field will not be updated.
 
-    :param organization_project_id: The ID of the github organization project.
-    :type organization_project_id: string/None
     :param organization_name: The github organization name.
     :type organization_name: string
+    :param organization_sfid: The SFDC identifier ID for the organization.
+    :type organization_sfid: string/None
     :param organization_installation_id: The github app installation id.
     :type organization_installation_id: string/None
     :return: dict representation of the new github organization object.
     :rtype: dict
     """
+
     github_organization = get_github_organization_instance()
     try:
         github_organization.load(str(organization_name))
     except DoesNotExist as err:
         return {'errors': {'repository_id': str(err)}}
     github_organization.set_organization_name(organization_name)
-    if organization_project_id:
-        github_organization.set_organization_company_id(str(organization_project_id))
     if organization_installation_id:
         github_organization.set_organization_installation_id(organization_installation_id)
+    if organization_sfid:
+        github_organization.set_organization_sfid(organization_sfid)    
     github_organization.save()
     return github_organization.to_dict()
 
 
-def delete_organization(organization_name):
+def delete_organization(auth_user, organization_name):
     """
     Deletes a github organization based on Name.
 
     :param organization_name: The Name of the github organization.
     :type organization_name: Name
     """
+    # Retrieve SFDC ID for this organization 
     github_organization = get_github_organization_instance()
     try:
         github_organization.load(str(organization_name))
     except DoesNotExist as err:
         return {'errors': {'organization_name': str(err)}}
+    
+    organization_sfid = github_organization.get_organization_sfid() 
+
+    # Validate user is authorized for this SFDC ID. 
+    can_access = check_user_authorization(auth_user, organization_sfid)
+    if not can_access['valid']:
+        return can_access['errors']
+
     github_organization.delete()
     return {'success': True}
 
@@ -130,17 +143,12 @@ def activity(body):
             existing = get_organization(body['installation']['account']['login'])
             if 'errors' in existing:
                 # TODO: Need a way of keeping track of new organizations that don't have projects yet.
-                org = create_organization(
-                    body['installation']['account']['login'],
-                    None,
-                    body['installation']['id']
-                )
-                return org
+                return {'status': 'Github Organization must be created through the Project Management Console.'}
             elif not existing['organization_installation_id']:
                 update_organization(
                     existing['organization_name'],
-                    existing['organization_company_id'],
-                    body['installation']['id']
+                    existing['organization_sfid'], 
+                    body['installation']['id'],
                 )
                 cla.log.info('Organization enrollment completed: %s', existing['organization_name'])
                 return {'status': 'Organization Enrollment Completed. CLA System is operational'}
